+++ pkg/deployment/members.go
@@ -30,6 +30,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/apis/shared"
 	"github.com/arangodb/kube-arangodb/pkg/deployment/reconcile"
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
 )
 
 func (d *Deployment) createAgencyMapping(ctx context.Context) error {
@@ -127,7 +128,13 @@ func (d *Deployment) renderMember(spec api.DeploymentSpec, status *api.Deploymen
 	}
 	deploymentName := apiObject.GetName()
 	role := group.AsRole()
+
 	arch := apiObject.GetAcceptedSpec().Architecture.GetDefault()
+	if arch != api.ArangoDeploymentArchitectureAMD64 && apiObject.Status.CurrentImage.ArangoDBVersion.CompareTo("3.10.0") < 0 {
+		arch = api.ArangoDeploymentArchitectureAMD64
+		d.log.Str("arch", string(arch)).Warn("Cannot render pod with requested arch. It's not supported in ArangoDB < 3.10.0. Defaulting architecture to AMD64")
+		d.CreateEvent(k8sutil.NewCannotSetArchitectureEvent(d.GetAPIObject(), string(arch), id))
+	}
 
 	switch group {
 	case api.ServerGroupSingle: