+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - (Dependencies) Bump K8S Dependencies to 1.22.15
 - (Bugfix) Unlock broken inspectors
 - (Debug) Allow to send package to stdout
+- (Improvement) ArangoDB image validation (=>3.10) for ARM64 architecture
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress